pyproject.toml
[tool.poetry]
name = "images-app"
version = "0.1.0"
description = ""
authors = ["irekpi <pilat.irek@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
pydantic = "^1.10.2"
sqlalchemy = "^1.4.45"
databases = {extras = ["sqlite"], version = "^0.7.0"}
python-multipart = "^0.0.5"
fastapi-pagination = "^0.11.1"
azure-storage-blob = "^12.14.1"
pillow = "^9.3.0"
pytest = "^7.2.0"
httpx = "^0.23.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
